+++ src/browser/index.ts
@@ -1,3 +1,4 @@
 /// <amd-module name='graphqlade/dist/browser'/>
+export * from "../util/AsyncPushIterator";
 export * from "../ws/GraphQLClientWebSocket";
 export * from "../ws/GraphQLWebSocketClient";